physics_rag_weaviate/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# AI/ML dependencies
google-generativeai>=0.3.0
numpy>=1.24.0

# Weaviate dependencies
weaviate-client>=4.6.0

# Search dependencies (fallback if needed)
rank-bm25>=0.2.2
nltk>=3.8.1

# Utility dependencies
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
